Initial functionality and testing
The contract allows the user to input the amount of eth to bet. Then the contract retrieves a random number generated from an oracle. If that number is 1, the user wins double their bet, if the number is 0 they lose their money. There was a problem discovered in testing where the oracle stops responding and the user is not able to get their money back nor bet further. I added functionality to check if the oracle has not responded in 10 minutes, if so the user gets their money back and they can try again. Further testing is needed to make sure this functionality works. Once that is confirmed, further functionality will be added so the user can enter a list of wallet addresses and the amount of eth to be raffled off. The contract will retrieve a random number from 0 to the number of addresses entered and the address that is stored at the index corresponding to the random number will win the eth.